+++ src/modules.js
@@ -137,10 +137,8 @@ var LibraryManager = {
       }
     }
 
-    // If there are any explicitly specified system JS libraries to link to, add those to link.
-    if (SYSTEM_JS_LIBRARIES) {
-      libraries = libraries.concat(SYSTEM_JS_LIBRARIES.split(','));
-    }
+    // Add any explicitly specified system JS libraries to link to, add those to link.
+    libraries = libraries.concat(SYSTEM_JS_LIBRARIES)
 
     if (LZ4) {
       libraries.push('library_lz4.js');
@@ -443,6 +441,7 @@ function exportRuntime() {
     'getTempRet0',
     'setTempRet0',
     'callMain',
+    'abort',
   ];
 
   if (!MINIMAL_RUNTIME) {